Add button to open GitHub edit
To submit changes to a blog article, – without this patch – you need to
find the repository, fork it, find the blog post change it and open a PR
to the main repository. This takes some time and discourages small fixes
like typos or small errors.

To address this, this patch adds a configurable edit button to the
bottom of every article. In the current configuration, clicking this
link will take you to the GitHub edit page for the article. If you did
not fork the repository yet, you get a button to fork it and then get
directly tossed into the editor to make changes. The changes can then
easily be opened as a PR from the same webpage.

There are a few options that can be configured in the `editPost`
parameter:
- `Text`: This is the text that will be displayed at the bottom of the
  article.
- `URL`: This will be the link the button leads to
- `appendFilePath`: If this is set to false, the target of the link will
  always be the same, e.g. only the repository's GitHub page. If it is
  set to true, the name of the article path will be appended.

These options enable flexibility should the repository ever move or if
you would like to point to somewhere else, like the GitHub IDE.
Removing `editPost` hash will disable the button entirely.